+++ src/modules/session/swagger/client/__init__.py
@@ -1,0 +1,13 @@
+"""Swagger client functionality."""
+
+from .base import SwaggerClient
+from .swagger2 import Swagger2Client
+from .openapi3 import OpenAPI3Client
+from .factory import SwaggerClientFactory
+
+__all__ = [
+    'SwaggerClient', 
+    'Swagger2Client', 
+    'OpenAPI3Client',
+    'SwaggerClientFactory'
+]
