+++ src/modules/session/swagger/client/base.py
