--- src/modules/swagger/__init__.py
@@ -1,5 +1,0 @@
-"""Swagger/OpenAPI integration for RestBook."""
-
-from .parser import SwaggerParser, SwaggerParserError
-
-__all__ = ["SwaggerParser", "SwaggerParserError"]
